Fix the logic for the full setup
In details, the full setup must *not* be called if the shell is
in the state (!login && interactive). This just means that a subshell
has been created and that the environment is already present.

In the script, the test is reverse. We just source the full environment
if:

  !(!login && interactive)

which we can also write:

  (login || ! interactive)

To be noted that for bash and tcsh, interactive means to have a prompt
(PS1 or prompt)