Add the detection of a custom root build (with --root)
This can be used to detect if the install script is being used inside a
rpm creation. In that case, the setup.py custom post-install should not
be triggered. 

The post-install action should be embedded in the SPEC file or with the
bdist_rpm options:
setup(
...
options = {'bdist_rpm':{'post_install' : 'post_install',
                        'post_uninstall' : 'post_uninstall'}},
...)
or
python setup.py bdist_rpm --post-install=<file>